Added shader reflection for dynamic pipeline layout creation.

+++ VirtualVistaVulkan/VulkanBuffer.cpp
@@ -97,7 +97,10 @@ namespace vv
         {
             buffer_create_info.sharingMode = VK_SHARING_MODE_CONCURRENT;
             buffer_create_info.queueFamilyIndexCount = 2;
-            std::array<uint32_t, 2> queue_family_indices = { device_->graphics_family_index, device_->transfer_family_index };
+            std::array<uint32_t, 2> queue_family_indices = {
+                static_cast<uint32_t>(device_->graphics_family_index),
+                static_cast<uint32_t>(device_->transfer_family_index)
+            };
             buffer_create_info.pQueueFamilyIndices = queue_family_indices.data();
         }
         else
